Tiny fix to e2e.sh (#876)
* Clean up CSS and make our typography a tiny bit nicer.

* Clean up colors in App.scss

* Fix font size in Chart and GraphVizChart

* Make inline `code` look green rathen than pink.

* Reduce h1 and h2 font sizes a bit.

* Tighten the spacing around lists and blockquotes.

* Fix e2e.sh so it doesn't autoexits when pgrep has an exit code

* Add comment to e2e.sh

* Change quotes in e2e.sh.

+++ scripts/e2e.sh
@@ -1,8 +1,9 @@
-#!/bin/bash -e
+#!/bin/sh
 ## Used to run end-to-end tests
 
 # Kill all active "streamlit run" processes
 pids=$(pgrep -f 'streamlit run')
+
 if [ "$pids" ]
 then
   kill -9 $pids
@@ -14,6 +15,12 @@ do
   # Run next test
   streamlit run $file &
   yarn --cwd "frontend" cy:run --spec "cypress/integration/${file%.*}.spec.ts"
+
+  # If exit code is nonzero, quit
+  if [ "$?" -ne "0"]
+  then
+    exit 1
+  fi
 
   # Kill the last process executed in the background
   kill -9 $!
